SmartLight Automatic and Manual
User-centric control strategy for home lighting system. Development of a
control strategy for managing the home lighting system taking into
account the user behaviours (e.g.. preferences and user-presence). The
inputs for the algorithm will be status of the appliance (on/off) and
user-presence (MQTT) events detected by IoT devices. The algorithm will
generate as output commands to manage the lighting systems.